Add nrf-softdevice feature
Problem as well as solution is documented in the [README][github] for
embassy's nrf-softdevice crate. Please refer to the content there for
details.

In short; The application code is sharing the cpu with the vendor
provided part of the firmware when using a nrf-softdevice. In order to
keep radio communications and such working, the three interrupts the
softdevice relies upon need not be blocked.

Section for a draft pull-request description
--------------------------------------------
In case there is disagreement on whether this is the correct crate for
this or not, it might be worth at least being more explicit in the README
about how this crate works by blocking all interrupts. Maybe one might
wish to also extend the API documentation mentioning this? We're open
for a conversation on how to best make this easy-to-overlook fact more
visible.

[github]: https://github.com/embassy-rs/nrf-softdevice#critical-sections